Move communications with main process to vuex store action
